% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allSquares.R
\name{allSquares}
\alias{allSquares}
\alias{allSquares,ANY-method}
\title{Calculate all squares}
\usage{
allSquares(x, y, ...)
}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}
}
\value{
An object of class AllSquares containing
 \item{addSquares}{The sum of the squared values}
 \item{subtractSquares}{The difference of the squared values}
 \item{x}{The first object input} 
 \item{y}{The second object input}
}
\description{
Calculates both added and subtracted squares
}
\note{
This produces an object of a new class
}
\examples{

myX <- c(20, 3) 
myY <- c(-2, 4.1) 
allSquares(myX, myY)
}
\seealso{
\code{\link{subtractSquares}}, \code{\link{addSquares}}
}
\author{
Jacob M. Montgomery
}
