% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/David.R
\docType{class}
\name{David-class}
\alias{AllSquares-class}
\alias{David-class}
\alias{getSquares,AllSquares-method}
\alias{initialize,AllSquares-method}
\title{A david value object}
\description{
Object of class \code{David} as created by the \code{David} functions
}
\details{
An object of the class `David' has the following slots:
\itemize{
\item \code{David} a numeric David
\item \code{x} The first input
\item \code{y} the second input 
}
}
\author{
David P. Flasterstein and Taishi T. Tamuroka: \email{jacob.montgomery@wustl.edu}
}

