% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTaishi.R
\name{printTaishi}
\alias{printTaishi}
\alias{printTaishi,ANY-method}
\title{Prints object Taishi}
\usage{
printTaishi(x, y, ...)
}
\arguments{
\item{Squares}{An object of class squares}
}
\value{
An object of class Squares containing
 \item{Taishi}{The word "Taishi"}
}
\description{
Prints objectTaishi on the screen
}
\note{
This is a very complex and misunderstood function
}
\examples{

myX <- 1
myY <- 2
printTaishi(myX, myY)
}
\author{
David P. Flasterstein and Taishi T. Muraoka
}
\seealso{
\code{\link{printDavid}}
}

