% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Squares.R, R/AllSquares-class.R
\docType{class}
\name{Squares-class}
\alias{Squares-class}
\alias{initialize,Squares-method}
\alias{getSquares,Squares-method}
\alias{getSquares}
\alias{AllSquares-class}
\alias{initialize,AllSquares-method}
\alias{getSquares,AllSquares-method}
\title{A squared value object}
\usage{
getSquares(object = "Squares")
}
\description{
Object of class \code{SquaresPack} are created by the \code{addSquares} and \code{subtractSquares} functions

Object of class \code{AllSquares} as created by the \code{allSquares} functions
}
\details{
An object of the class `SquaresPack' has the following slots:
\itemize{
\item \code{square} The added or subtracted squared values
\item \code{x} The first input
\item \code{y} the second input 
}

An object of the class `Squares' has the following slots:
\itemize{
\item \code{addSquare} The added or subtracted squared values
\item \code{subtractSquare} The added or subtracted squared values
\item \code{x} The first input
\item \code{y} the second input 
}
}
\author{
Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}

Jacob M. Montgomery: \email{jacob.montgomery@wustl.edu}
}
