% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractSquares.R
\name{subtractSquares}
\alias{subtractSquares}
\title{Subtracting squared values}
\usage{
subtractSquares(x, y)
}
\arguments{
\item{x}{A numeric object}

\item{y}{A numeric object with the same dimensionality as \code{x}.}
}
\value{
A list with the elements
 \item{squares}{The sum of the squared values}
 \item{x}{The first object input} 
 \item{y}{The second object input}
}
\description{
Finds the difference of squared of numbers
}
\note{
This is a very simple function
}
\examples{

myX <- c(20, 3) 
myY <- c(-2, 4.1) 
subtractSquares(myX, myY)
}
\seealso{
addSquares
}
\author{
Jacob M. Montgomery<\email{jacob.montgomery@wustl.edu}>
}
