% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSquares.R
\name{addSquares}
\alias{addSquares}
\title{Function that adds squares}
\usage{
addSquares(x, y)
}
\arguments{
\item{x}{Numeric}

\item{y}{Numeric}
}
\value{
List of 3 elements.
}
\description{
Function that adds squares
}
