\docType{package}
\name{squaresPack}
\alias{squaresPack}
\alias{squaresPack-package}
\title{squaresPack}
\description{
The squaresPack package performs simple arithmetic
calculations.
}
\examples{
\dontrun{
demo(addSquares)
demo(subtractSquares)
}
}
\author{
Ryan T. Moore: \email{rtm@wustl.edu} and Jacob M.
Montgomery: \email{jacob.montgomery@wustl.edu}
}

