% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractSquares.R
\name{subtractSquares}
\alias{subtractSquares}
\title{Subtracts squares}
\usage{
subtractSquares(x, y)
}
\arguments{
\item{x}{Numeric}

\item{y}{Numeric}
}
\value{
List of three elements.
}
\description{
Subtracts squares
}
